+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [Unrealeased]
+## [To Add]
+- Dynamic Graph generator
+- Dynamic Graph upload/load
+
+## [2.0.0]
 ### Added
 - Python 3 support
+- Dynamic Network Support
+	- Models
+		- Dynamic SI
+		- Dynamic SIS
+		- Dynamic SIR
      
 ## [1.0.0]
 ### Added